Centralize iframe constants in iframe-utils
Address ce:review findings:
- Move UNSUPPORTED_IFRAME_SRCS, frame-depth clamping, and the
  ignore-selector resolver into iframe-utils.js so the same constants
  don't drift across SDKs.
- isUnsupportedIframeSrc lowercases the src so JavaScript:/DATA:
  prefixes are caught.
